chore: fix all compiler warnings and pass ci checks
- suppress dead_code warnings with allow attributes for unused but intentional code
- mark test-only methods with cfg(test) (time_since_flush, get_last_processed_block)
- fix clippy errors: remove unnecessary cast in transform/mod.rs
- remove always-true assertion in exex/mod.rs
- remove unused imports in tests and modules
- fix example import path for clickhousewriter
- run cargo fmt --all for consistent formatting
- ignore python __pycache__ directories

all ci checks passing:
- cargo fmt --all -- --check: PASS
- cargo clippy --all-targets --all-features -- -D warnings: PASS
- cargo test --lib: 42 tests passed

+++ src/transform/mod.rs
@@ -57,7 +57,7 @@ fn transform_transaction(
 
     let gas_limit = tx.gas_limit();
 
-    let gas_price = tx.gas_price().unwrap_or(0) as u128;
+    let gas_price = tx.gas_price().unwrap_or(0);
 
     let max_fee_per_gas = Some(tx.max_fee_per_gas());
     let max_priority_fee_per_gas = tx.max_priority_fee_per_gas();
